requirements.txt
python-telegram-bot
google-generativeai
groq
notion-client
python-dotenv
flask
requests
apscheduler